Nose should ignore this test
