+++ yt/geometry/tests/test_particle_octree.py
@@ -6,13 +6,10 @@ import yt.units.dimensions as dimensions
 from yt.geometry.oct_container import _ORDER_MAX
 from yt.geometry.particle_oct_container import ParticleBitmap, ParticleOctreeContainer
 from yt.geometry.selection_routines import RegionSelector
-from yt.loaders import load
-from yt.sample_data.api import _get_test_data_dir_path
 from yt.testing import (
     assert_array_equal,
     assert_equal,
     assert_true,
-    requires_file,
     requires_module,
 )
 from yt.units.unit_registry import UnitRegistry
@@ -404,35 +401,6 @@ def test_bitmap_select():
                         gf_ans.append(rf_ghost % nfiles)
                     gf_ans = np.array(sorted(gf_ans))
                     assert_array_equal(gf, gf_ans, f"selector {i}, ghost file array")
-
-
-@requires_file("TNGHalo/halo_59.hdf5")
-def test_ewah_write_load(tmp_path):
-    mock_file = tmp_path / "halo_59.hdf5"
-    mock_file.symlink_to(_get_test_data_dir_path() / "TNGHalo" / "halo_59.hdf5")
-
-    masses = []
-    opts = [
-        (None, True, (6, 2, 4), False),
-        (None, True, (6, 2, 4), True),
-        ((5, 3), False, (5, 3, 3), False),
-        ((5, 3), False, (5, 3, 3), True),
-    ]
-
-    for opt in opts:
-        ds = load(mock_file, index_order=opt[0])
-        _, c = ds.find_max(("gas", "density"))
-        assert ds.index.pii.mutable_index is opt[1]
-        assert ds.index.pii.order1 == opt[2][0]
-        assert ds.index.pii.order2_orig == opt[2][1]
-        assert ds.index.pii.order2 == opt[2][2]
-        assert ds.index.pii._is_loaded is opt[3]
-        c += ds.quan(0.5, "Mpc")
-        sp = ds.sphere(c, (20.0, "kpc"))
-        mass = sp.sum(("gas", "mass"))
-        masses.append(mass)
-
-    assert_array_equal(masses[0], masses[1:])
 
 
 def cell_centers(order, left_edge, right_edge):
